Moved all expr's into expr_visitor and stmt helper methods into stmt_visitor
